requirements.txt
pymongo
requests
pandas
numpy
scikit-learn
joblib
python-dotenv
tensorflow
xgboost
keras
